Refactor tests to Role-Centric architecture
Moved tests from `tests/config` and `tests/ansible` to `tests/roles/<role_name>`
and `tests/intg`. This aligns the test structure with the "Structure Mirroring"
principle, consolidating tests related to each role.

Changes:
- Created `tests/roles/` for role-specific tests.
- Created `tests/intg/` for integration tests (formerly `tests/ansible`).
- Renamed and moved test files to their respective role directories.
- Added `__init__.py` to role test directories to avoid module name collisions.
- Updated `tests/intg/test_role_integrity.py` imports.
- Updated `tests/intg/test_checksums.py` file path reference.
- Moved `tests/config/conftest.py` to `tests/roles/editor/conftest.py`.
