--- tests/config/rust/__init__.py
@@ -1,3 +1,0 @@
-"""Tests for Rust configuration assets."""
-
-__all__ = []
