+++ tests/intg/installer_checksums.yml
