+++ tests/intg/test_justfile.py
