+++ tests/roles/editor/test_config.py
