+++ tests/roles/shell/test_aider_cli.py
