+++ tests/roles/shell/test_gemini_gen.py
