+++ tests/roles/shell/test_ssh_manager.py
